removed unused code
